Pivotal ID # 170993540: Refactor Processors To Services
- Remove unnecessary files
- Improve classes naming
