--- submission/submitter/src/main/kotlin/ac/uk/ebi/biostd/submission/processors/AccNoProcessor.kt
@@ -1,74 +1,0 @@
-package ac.uk.ebi.biostd.submission.processors
-
-import ac.uk.ebi.biostd.submission.exceptions.ProvideAccessNumber
-import ac.uk.ebi.biostd.submission.exceptions.UserCanNotUpdateSubmit
-import ac.uk.ebi.biostd.submission.util.AccNoPatternUtil
-import arrow.core.Option
-import arrow.core.getOrElse
-import ebi.ac.uk.base.lastDigits
-import ebi.ac.uk.model.AccNumber
-import ebi.ac.uk.model.ExtendedSubmission
-import ebi.ac.uk.model.extensions.attachTo
-import ebi.ac.uk.persistence.PersistenceContext
-import ebi.ac.uk.security.integration.components.IUserPrivilegesService
-
-const val DEFAULT_PATTERN = "!{S-BSST}"
-const val PATH_DIGITS = 3
-
-/**
- * Calculate the accession number and relative path for the given submission.
- */
-class AccNoProcessor(
-    private val userPrivilegesService: IUserPrivilegesService,
-    private val patternUtil: AccNoPatternUtil
-) : SubmissionProcessor {
-    override fun process(submission: ExtendedSubmission, context: PersistenceContext) {
-        val accNo = getAccNo(submission, context)
-
-        submission.accNo = accNo.toString()
-        submission.relPath = getRelPath(accNo)
-    }
-
-    private fun getAccNo(submission: ExtendedSubmission, context: PersistenceContext): AccNumber {
-        return when {
-            submission.accNo.isNotEmpty() &&
-                context.isNew(submission.accNo) &&
-                userPrivilegesService.canProvideAccNo(submission.user.email).not() ->
-                throw ProvideAccessNumber(submission.user)
-            userPrivilegesService.canResubmit(
-                submission.user.email, submission.user, submission.attachTo, submission.accessTags).not() ->
-                throw UserCanNotUpdateSubmit(submission)
-            else ->
-                calculateAccNo(submission, context)
-        }
-    }
-
-    private fun calculateAccNo(submission: ExtendedSubmission, context: PersistenceContext): AccNumber {
-        return when {
-            submission.accNo.isEmpty() ->
-                calculateAccNo(getPatternOrDefault(context.getParentAccPattern(submission)), context)
-            patternUtil.isPattern(submission.accNo) ->
-                calculateAccNo(patternUtil.getPattern(submission.accNo), context)
-            else -> patternUtil.toAccNumber(submission.accNo)
-        }
-    }
-
-    private fun calculateAccNo(prefix: String, context: PersistenceContext) =
-        AccNumber(prefix, context.getSequenceNextValue(prefix))
-
-    @Suppress("MagicNumber")
-    internal fun getRelPath(accNo: AccNumber): String {
-        val prefix = accNo.prefix
-        val value = accNo.numericValue
-
-        return when {
-            accNo.numericValue < 99 ->
-                "$prefix/${prefix}0-99/$prefix$value".removePrefix("/")
-            else ->
-                "$prefix/${prefix}xxx${value.lastDigits(PATH_DIGITS)}/$prefix$value".removePrefix("/")
-        }
-    }
-
-    private fun getPatternOrDefault(pattern: Option<String>) =
-        pattern.map { patternUtil.getPattern(it) }.getOrElse { patternUtil.getPattern(DEFAULT_PATTERN) }
-}
