--- submission/submitter/src/main/kotlin/ac/uk/ebi/biostd/submission/processors/AccessTagProcessor.kt
@@ -1,14 +1,0 @@
-package ac.uk.ebi.biostd.submission.processors
-
-import ebi.ac.uk.model.ExtendedSubmission
-import ebi.ac.uk.persistence.PersistenceContext
-
-/**
- * Add all parent submission tags to submitted submission.
- */
-class AccessTagProcessor : SubmissionProcessor {
-    override fun process(submission: ExtendedSubmission, context: PersistenceContext) {
-        val accessTags = context.getParentAccessTags(submission).filterNot { it == "Public" }
-        submission.accessTags.addAll(accessTags)
-    }
-}
