--- submission/submitter/src/main/kotlin/ac/uk/ebi/biostd/submission/processors/SubmissionProcessor.kt
@@ -1,8 +1,0 @@
-package ac.uk.ebi.biostd.submission.processors
-
-import ebi.ac.uk.model.ExtendedSubmission
-import ebi.ac.uk.persistence.PersistenceContext
-
-interface SubmissionProcessor {
-    fun process(submission: ExtendedSubmission, context: PersistenceContext)
-}
